pyproject.toml
[tool.poetry]
name = "labgas-manager"
version = "0.1.0"
description = "Dashboard para gestão de cilindro e elementos em laboratório de química"
authors = ["Lucas Cavalcante dos Santos <cavalcanteprofissional@outlook.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.28"
pandas = "^2.0"
plotly = "^5.17"
supabase = "^1.2"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0"
ruff = "^0.1"
pytest = "^7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
